+++ frontend/src/layouts/RootLayout.jsx
@@ -3,6 +3,7 @@ import { Outlet, useNavigation } from 'react-router'
 import Header from "../components/Header.jsx";
 import Footer from "../components/Footer.jsx";
 import Loading from "../components/Loading.jsx";
+import WhatsAppCircle from "../components/WhatsAppCircle.jsx";
 
 export default function RootLayout() {
   const navigation = useNavigation();
@@ -17,6 +18,7 @@ export default function RootLayout() {
       <Header />
       <Outlet/>
       <Footer />
+      <WhatsAppCircle/>
       </>
     )}
   </>